% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Draw a name}
\usage{
draw_a_name(pren = "Vincent", s = "M")
}
\arguments{
\item{pren}{the name you are analysing}

\item{s}{the sex}
}
\value{
the chart of two names
}
\description{
Draw a name
}
