% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Draw names}
\usage{
draw_names(pren = c("Pierre", "Paul"))
}
\arguments{
\item{pren}{the two names which will be plotted defaulted to Pierre and Paul}
}
\value{
draws the chart representing the historic trend of two names
}
\description{
Draw names
}
