% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_excel.R
\name{multi_excel}
\alias{multi_excel}
\title{Read multiple excel files}
\usage{
multi_excel(file)
}
\arguments{
\item{file}{file with excel sheets}
}
\value{
a list
}
\description{
Read multiple excel files
}
