% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prenoms_dataset.R
\docType{data}
\name{prenoms}
\alias{prenoms}
\title{Prenoms Dataset}
\description{
French baby names between 1900 and 2015, detailed by department
}
\references{
\url{https://www.insee.fr/fr/statistiques/2540004}{a link}
}
\author{
INSEE
}
\keyword{data}
