% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_as_csv.R
\name{save_as_csv}
\alias{save_as_csv}
\title{Save csv files}
\usage{
save_as_csv(x, file_name)
}
\arguments{
\item{x}{path to the folder}

\item{file_name}{name of the file}
}
\value{
file name
}
\description{
Save csv files
}
